riscv: sys_io: move custom sys_io to be generic
84b86d9b0 introduced the ability of defining custom sys_io functions for
riscv. This is however useful for all other architectures and there is
nothing riscv specific about it. So instead of having this be riscv
specific, move implementation to the common architecture code and make
it available for everyone using PLATFORM_HAS_CUSTOM_SYS_IO.
